feat: Add SVG icons for Vite and React
This commit adds SVG icons for Vite and React to the web-react project. The Vite icon is added as a new file in the `public` directory, while the React icon is added as a new file in the `src/assets` directory. These icons will be used to enhance the visual representation of the respective technologies in the application.