feat: Add Button and Input components to web-react
This commit adds the Button and Input components to the web-react project. The Button component is created in the `src/components/button.tsx` file and includes variants for primary and secondary styles, as well as support for loading and disabled states. The Input component is created in the `src/components/input.tsx` file and includes support for labels, icons, and different input types. These components will be used for creating interactive forms and buttons throughout the application.