feat: Add IPerson interface to web-react
This commit adds the IPerson interface to the web-react project. The `IPerson.ts` file is created in the `src/interface` directory and includes properties for representing a person's information such as cpf, name, rg, zipCode, address, number, neighborhood, complement, city, and state. This interface will be used for defining the structure of person objects throughout the application.