feat: Update getPerson route to search by cpf
The getPerson route in the back-end server is updated to search for a person's information using their cpf (Cadastro de Pessoas Físicas). The `get-person.ts` file is modified to include a query parameter for cpf and use the `contains` operator to find matching records. This change improves the functionality of the route and allows for more flexible search options.